Extracted ItemAdapter to a separate class
