+++ app/src/main/res/values-v21/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="rouble">\u20bd</string>
+</resources>
