+++ app/src/main/res/values/strings.xml
@@ -3,5 +3,5 @@
     <string name="name_hint">Название</string>
     <string name="money_sign">Р</string>
     <string name="amount_hint">0</string>
-    <string name="rouble">\u20bd</string>
+    <string name="rouble">Р</string>
 </resources>
